Main: selections are a tuple of (event, ordinal)
where ordinal is occurrence.ordinal of the occurrence under the mouse.
this is the only way to reliably keep track of which instance of an event
is selected, since the occurrences cache every time the calendar
is re-drawn and all new recurrence objects are created every time an
event changes.